requirements.txt
asgiref==3.8.1
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
environ==1.0
Markdown==3.8
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-environ==0.4.54
sqlparse==0.5.3
